+++ client/galaxy/scripts/components/Workflow/Editor/WorkflowRecommendations.vue
@@ -43,9 +43,6 @@ export default {
     props: {
         workflowManager: {
             type: Object
-        },
-        toolSequence: {
-            type: String
         }
     },
     data() {
@@ -65,10 +62,64 @@ export default {
     created() {
         this.loadRecommendations();
     },
-    computed: {},
     methods: {
+        getToolId(toolId) {
+            if (toolId !== undefined && toolId !== null && toolId.indexOf("/") > -1) {
+                const toolIdSlash = toolId.split("/");
+                toolId = toolIdSlash[toolIdSlash.length - 2];
+            }
+            return toolId;
+        },
+        getWorkflowPath(wfSteps, currentNodeId) {
+            const steps = {};
+            const stepNames = {};
+            for (const stpIdx in wfSteps.steps) {
+                const step = wfSteps.steps[stpIdx];
+                const inputConnections = step.input_connections;
+                stepNames[step.id] = this.getToolId(step.content_id);
+                for (const icIdx in inputConnections) {
+                    const ic = inputConnections[icIdx];
+                    if (ic !== null && ic !== undefined) {
+                        const prevConn = [];
+                        for (const conn of ic) {
+                            prevConn.push(conn.id.toString());
+                        }
+                        steps[step.id.toString()] = prevConn;
+                    }
+                }
+            }
+            // recursive call to determine path
+            function readPaths(nodeId, ph) {
+                for (const st in steps) {
+                    if (parseInt(st) === parseInt(nodeId)) {
+                        const parentId = parseInt(steps[st][0]);
+                        if (parentId !== undefined && parentId !== null) {
+                            ph.push(parentId);
+                            if (steps[parentId] !== undefined && steps[parentId] !== null) {
+                                readPaths(parentId, ph);
+                            }
+                        }
+                    }
+                }
+                return ph;
+            }
+            let ph = [];
+            const stepNameList = [];
+            ph.push(currentNodeId);
+            ph = readPaths(currentNodeId, ph);
+            for (const sIdx of ph) {
+                const sName = stepNames[sIdx.toString()];
+                if (sName !== undefined && sName !== null) {
+                    stepNameList.push(sName);
+                }
+            }
+            return stepNameList.join(",");
+        },
         loadRecommendations() {
-            const requestData = { tool_sequence: this.toolSequence };
+            const workflowSimple = this.workflowManager.node.app.to_simple();
+            const node = this.workflowManager.node;
+            const toolSequence = this.getWorkflowPath(workflowSimple, node.id);
+            const requestData = { tool_sequence: toolSequence };
             getToolPredictions(requestData).then(responsePred => {
                 const predictedData = responsePred.predicted_data;
                 const outputDatatypes = predictedData.o_extensions;
